Use chai.expect to test status codes
This fixes a problem that was introduced when I started using
the `end` method from supertest. If you are using the `end` method
`expect` assertions that fail will not throw - they will return
the assertion as an error to the `end` callback. In order to fail
the test case, you will need to rethrow or pass `err` to `done` if
it is defined.